fix: fix error module not found after refactor
