+++ src/App.js
@@ -7,15 +7,15 @@ import AddData from "./pages/AddData";
 import Schedule from "./pages/Schedule";
 import EditProfile from "./pages/EditProfile";
 import Employee from "./pages/Employee";
-import BlogPost from "./BlogPost/pages/BlogPost";
-import EditBlogPost from "./BlogPost/pages/EditBlogPost";
+import BlogPost from "./V2/BlogPost/pages/BlogPost";
+import EditBlogPost from "./V2/BlogPost/pages/EditBlogPost";
 import AddBlogPost from "./pages/AddBlogPost";
 import AddEmployee from "./pages/AddEmployee";
 import EmpAdhoc from "./V2/EmpAdhoc/pages/EmpAdhoc";
 // import Home from "./pages/guest/Home";
 import PrintEmployeeAdhoc from "./pages/PrintEmployeeAdhoc";
 import { AdminRoute, GuestRoute } from "./middleware/AuthRoute";
-import CreateBlogPost from "./BlogPost/pages/CreateBlogPost";
+import CreateBlogPost from "./V2/BlogPost/pages/CreateBlogPost";
 import LandingPage from "../src/LandingPage"
 
 function App() {
